integrations/google-workspace-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "google-workspace-mcp"
version = "1.0.0"
description = "MCP server for Google Workspace (Sheets, Drive, Gmail)"
authors = [{name = "Brickface", email = "fred@brickface.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
]

[project.scripts]
google-workspace-mcp = "google_sheets_mcp:main"
